Refactored celery configuration
