+++ app/__init__.py
@@ -1,6 +1,6 @@
 import logging
-from app.main import app
+from app.main import create_app
 
-notify_orchestrator_api = app
-celery = app.celery_app
+notify_orchestrator_api = create_app()
+celery = notify_orchestrator_api.celery_app
 logger = logging.getLogger("uvicorn")
