--- config/celery_config.py
@@ -1,38 +1,0 @@
-import os
-from functools import lru_cache
-from kombu import Queue
-
-
-def route_task(name, args, kwargs, options, task=None, **kw):
-    if ":" in name:
-        queue, _ = name.split(":")
-        return {"queue": queue}
-    return {"queue": "celery"}
-
-
-class BaseConfig:
-    CELERY_BROKER_URL: str = os.environ.get("CELERY_BROKER_URL", "amqp://guest:guest@localhost:5672//")
-
-    CELERY_TASK_QUEUES: list = (
-        # default queue
-        Queue("celery"),
-    )
-
-    CELERY_TASK_ROUTES = (route_task,)
-
-
-class DevelopmentConfig(BaseConfig):
-    pass
-
-
-@lru_cache()
-def get_settings():
-    config_cls_dict = {
-        "development": DevelopmentConfig,
-    }
-    config_name = os.environ.get("CELERY_CONFIG", "development")
-    config_cls = config_cls_dict[config_name]
-    return config_cls()
-
-
-settings = get_settings()
