+++ utils/celery.py
@@ -1,0 +1,21 @@
+from celery.result import AsyncResult
+
+
+def get_task_info(task_id):
+    """
+    return task info for the given task_id
+    """
+    task_result = AsyncResult(task_id)
+    result = {
+        "task_id": task_id,
+        "task_status": task_result.status,
+        "task_result": task_result.result
+    }
+    return result
+
+
+def route_task(name):
+    if ":" in name:
+        queue, _ = name.split(":")
+        return {"queue": queue}
+    return {"queue": "celery"}
